refactoring: 리뷰 수정 서버 연결 코드 viewModel로 이전
