+++ app/src/main/java/com/eatssu/android/ui/review/etc/FixViewModel.kt
@@ -6,6 +6,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.eatssu.android.data.service.ReviewService
+import com.eatssu.android.util.RetrofitImpl
 import com.eatssu.android.util.RetrofitImpl.retrofit
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -23,6 +24,37 @@ class FixViewModel : ViewModel() {
     private val _toastMessage = MutableLiveData<String>()
     val toastMessage: LiveData<String> get() = _toastMessage
 
+    fun postData(reviewId: Long, comment: String, mainGrade: Int, amountGrade: Int, tasteGrade: Int) {
+        val service = RetrofitImpl.retrofit.create(ReviewService::class.java)
+
+        val reviewData = """
+            {
+                "mainGrade": $mainGrade,
+                "amountGrade": $amountGrade,
+                "tasteGrade": $tasteGrade,
+                "content": "$comment"
+            }
+        """.trimIndent().toRequestBody("application/json".toMediaTypeOrNull())
+
+        viewModelScope.launch(Dispatchers.IO) {
+            service.modifyReview(reviewId, reviewData).enqueue(object : Callback<Void> {
+                override fun onResponse(call: Call<Void>, response: Response<Void>) {
+                    if (response.isSuccessful) {
+                        if (response.code() == 200) {
+                            handleSuccessResponse("수정이 완료되었습니다.")
+                        } else {
+                            handleErrorResponse("수정이 실패하였습니다.")
+                        }
+                    }
+                }
+
+                override fun onFailure(call: Call<Void>, t: Throwable) {
+                    handleErrorResponse("수정이 실패하였습니다.")
+                }
+            })
+        }
+    }
+
     fun handleSuccessResponse(message: String) {
         viewModelScope.launch(Dispatchers.Main) {
             _toastMessage.value = message
@@ -30,8 +62,6 @@ class FixViewModel : ViewModel() {
 
         }
     }
-
-
 
     fun handleErrorResponse(message: String) {
         viewModelScope.launch(Dispatchers.Main) {
